solve a nast bug related with _set, problems it takes integers instead of float

+++ object.c
@@ -55,7 +55,7 @@ GLubyte _indices[]	= {  0, 1, 2, 2, 3, 0};
 
 void _Render( RECTANGLE *, float, float);
 void _RenderInstances( RECTANGLE *, float, float, int);
-void _set( RECTANGLE *, int x, int y);
+void _set( RECTANGLE *, float x, float y);
 
 void Rectangle_Init( RECTANGLE * Rect, float width, float height, float x, float y){
 
@@ -87,7 +87,7 @@ void _RenderInstances( RECTANGLE * Rect, float x, float y, int instances){
 }
 
 
-void _set( RECTANGLE * Rect, int x, int y){
+void _set( RECTANGLE * Rect, float x, float y){
     Rect->vertices[0]  = x + Rect->width;   Rect->vertices[3]  = x - Rect->width;   Rect->vertices[6]  = x - Rect->width;   Rect->vertices[9]  = x + Rect->width;
     Rect->vertices[1]  = y + Rect->height;  Rect->vertices[4]  = y + Rect->height;  Rect->vertices[7]  = y - Rect->height;  Rect->vertices[10] = y - Rect->height;
     Rect->vertices[2]  = 0.0f;              Rect->vertices[5]  = 0.0f;              Rect->vertices[8]  = 0.0f;              Rect->vertices[11] = 0.0f;
